Add register peer tx
