+++ common/cosmwasm-smart-contracts/ephemera/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod msg;
+pub mod types;
