+++ common/cosmwasm-smart-contracts/ephemera/src/types.rs
@@ -1,0 +1,39 @@
+// Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, JsonSchema)]
+pub struct JsonPeerInfo {
+    /// The name of the peer. Can be arbitrary.
+    pub name: String,
+    /// The address of the peer. See [PeerInfo] for more details.
+    pub address: String,
+    ///Serialized public key.
+    ///
+    /// # Converting to string and back example
+    ///```
+    /// use ephemera::crypto::{EphemeraKeypair, EphemeraPublicKey, Keypair, PublicKey};
+    ///
+    /// let public_key = Keypair::generate(None).public_key();
+    ///
+    /// let public_key_str = public_key.to_string();
+    ///
+    /// let public_key_parsed = public_key_str.parse::<PublicKey>().unwrap();
+    ///
+    /// assert_eq!(public_key, public_key_parsed);
+    /// ```
+    pub public_key: String,
+}
+
+impl JsonPeerInfo {
+    #[must_use]
+    pub fn new(name: String, address: String, public_key: String) -> Self {
+        Self {
+            name,
+            address,
+            public_key,
+        }
+    }
+}
